add resource layer
