+++ packages/esm-patient-registration-app/src/patient-registration/input/comboinput/comboinput.component.tsx
@@ -1,9 +1,7 @@
-import { openmrsFetch } from '@openmrs/esm-framework';
 import { ComboBox, ComboBoxProps } from 'carbon-components-react';
 import { useField } from 'formik';
 import React, { useContext, useState } from 'react';
-import { PatientRegistrationContext } from '../../patient-registration-context';
-const AH_BASE_WS_API_URL = '/module/addresshierarchy/ajax/getPossibleAddressHierarchyEntriesWithParents.form';
+import { performAdressHirarchiWithParentSearch } from '../../../resource';
 
 interface InputProps extends ComboBoxProps {
   name: string;
@@ -11,14 +9,7 @@ interface InputProps extends ComboBoxProps {
   setSelectedValue: any;
   selected: string;
 }
-export function performAdressHirarchiWithParentSearch(addressField, parentid, query) {
-  return openmrsFetch(
-    `${AH_BASE_WS_API_URL}?addressField=${addressField}&limit=20&searchString=${query}&parentUuid=${parentid}`,
-    {
-      method: 'GET',
-    },
-  );
-}
+
 export const ComboInput: React.FC<InputProps> = ({ name, labeltext, setSelectedValue, selected }) => {
   const [field, Meta, helpers] = useField(name);
   const nulldata = [];
