+++ packages/esm-patient-registration-app/translations/en.json
@@ -56,7 +56,6 @@
   "restoreRelationshipActionButton": "Undo",
   "sexFieldLabelText": "Sex",
   "stateProvince": "Region",
-  "countyDistrict": "District",
   "stroke": "Stroke",
   "unidentifiedPatient": "Unidentified Patient",
   "unknown": "Unknown",
